- moves the authentication configuration to its own extension method
